html/semantics/forms/the-fieldset-element/HTMLFieldSetElement.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: HTMLFieldSetElement interface</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/the-button-element.html#the-fieldset-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <form name="fm1" style="display:none">
      <fieldset id="fs_outer">
        <legend><input name="txt1"></legend>
        <output name="op"></output>
        <button name="btn">BUTTON</button>
        <keygen name="kg"/>
        <select name="slt"></select>
        <object name="obj"/>
        <fieldset id="fs_inner">
          <textarea name="ta"></textarea>
        </fieldset>
      </fieldset>
    </form>
    <script>
      var fm1,
          fs_outer,
          cts_outer;

      setup(function () {
        fm1 = document.forms.fm1;
        fs_outer = document.getElementById("fs_outer");
        cts_outer = fs_outer.elements;
      });

      test(function () {
        assert_equals(fs_outer.type, "fieldset", "The value of type attribute is incorrect.");
      }, "The type attribute must return 'fieldset'");

      test(function () {
        assert_equals(fs_outer.form, fm1, "The fieldset should has a form owner.");
      }, "The form attribute must return the fieldset's form owner");

      test(function () {
        assert_true(cts_outer instanceof HTMLFormControlsCollection,
                    "The elements attribue should be an HTMLFormControlsCollection object");
      }, "The elements must retrun an HTMLFormControlsCollection object");

      test(function () {
        var fs_inner = document.getElementById("fs_inner");
        var cts_inner = fs_inner.elements;
        assert_array_equals(cts_inner, [fm1.ta],
                            "The items in the conllection must be children of the inner fieldset element.");
        assert_array_equals(cts_outer, [fm1.txt1, fm1.op, fm1.btn, fm1.kg, fm1.slt, fm1.obj, fm1.fs_inner, fm1.ta],
                            "The items in the conllection must be children of the outer fieldset element.");
      }, "The controls must root at the fieldset element");

      test(function () {
        assert_false(fs_outer.disabled, "The disabled attribute should be false.");
        fs_outer.setAttribute("disabled", "true");
        assert_true(fs_outer.disabled, "The disabled attribute should be true.");
      }, "The disabled attribute must reflect the content attribute");
    </script>
  </body>
</html>
